backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pillow>=10.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
mediapipe>=0.10.0
yt-dlp>=2023.0.0
